Hooke's law attenuation related to edges per node
Example: you have 130 [0 ... 129]. Each node from 1 to 129
share an edge with node 0. Thus, node 0 has 129 edges, the
edges [1 ... 129] have one edge for node. Without attenuation
of the Hooke's law, the springs' length for node 0 increase
dramatically and the graph explode producing the following
situation:

  node_0 (out of canvas)
        +-----------------------+
        |\                      |
        |  \         [canvas]   |
        |    \                  |
        |      \                |
        |        \\             |
        |          \\ edges     |
        |             \\        |
        |                \\     |
        +-----------------------+
                    nodes_[1 ... 129] (out of canvas)

this is because hooke's law is applied 130 times for a single
node on every of its edges. So, we use edge's count for every node
as a divisor for the calculated force.